Factored out the yelp logic into its own controller
